requirements.txt
aiohappyeyeballs==2.4.3
aiohttp==3.10.10
aiosignal==1.3.1
async-timeout==4.0.3
attrs==24.2.0
blinker==1.9.0
certifi==2024.8.30
charset-normalizer==3.4.0
click==8.1.8
colorama==0.4.6
conda-pack @ file:///home/conda/feedstock_root/build_artifacts/conda-pack_1731874358732/work
dash==3.2.0
efficientnet_pytorch==0.7.1
et_xmlfile==2.0.0
filelock==3.16.1
Flask==3.1.2
frozenlist==1.5.0
fsspec==2024.10.0
huggingface-hub==0.26.2
idna==3.10
imageio==2.36.0
importlib_metadata==8.7.0
intel-openmp==2021.4.0
itsdangerous==2.2.0
Jinja2==3.1.4
joblib==1.5.2
lazy_loader==0.4
lightning-utilities==0.11.8
MarkupSafe==3.0.2
mkl==2021.4.0
mpmath==1.3.0
multidict==6.1.0
munch==4.0.0
narwhals==2.6.0
nest-asyncio==1.6.0
networkx==3.2.1
numpy==1.26.3
openpyxl==3.1.5
packaging==24.2
pandas==2.3.3
pillow==11.0.0
plotly==6.3.0
pretrainedmodels==0.7.4
propcache==0.2.0
python-dateutil==2.9.0.post0
pytorch-lightning==2.3.0
pytz==2025.2
PyYAML==6.0.2
requests==2.32.3
retrying==1.4.2
safetensors==0.4.5
scikit-image==0.24.0
scikit-learn==1.2.2
scipy==1.13.1
segmentation-models-pytorch==0.3.3
six==1.16.0
sympy==1.13.1
tbb==2021.11.0
threadpoolctl==3.6.0
tifffile==2024.8.30
timm==0.9.2
torch==2.3.1+cu118
torchmetrics==1.5.2
torchvision==0.18.1+cu118
tqdm==4.66.4
typing_extensions==4.12.2
tzdata==2025.2
urllib3==2.2.3
Werkzeug==3.1.3
yarl==1.17.1
zipp==3.23.0
